Initial commit of Queue Album Plugin
This sets up a plugin that allows you to add all songs in the
album currently selected to the play queue, allowing you to build a
one-time only playlist of albums quickly and easily.  The current
implementation isn't ideal (it loops instead of using DB queries)
and it simply chooses to ignore selections that include multiple
entries. In addition, the menu item really should be in the context
menu for entries, but I have no idea how to set that up (and the docs
are no help). But it works as a first cut.